Index the source file for each symbol
* bdx/binary.py (Symbol): Add `source' attribute.
(CompilationDatabase): New class.
(_read_compdb):
(_find_source_file_compdb):
(_find_source_file_dwarfdump):
(_find_source_file): New.
(_read_symbols_in_file): Rename from `_read_symtab'.  Set the new
`source' attribute.
(read_symbols_in_file): Rename from `read_symtable'.
(BinaryDirectory._find_files_from_compilation_database): Use
`_read_compdb'.
* bdx/cli.py (sexp_format): Output None values as `nil'.
* bdx/index.py (IndexingOptions): Add `use_dwarfdump' attribute.
(PathField.preprocess_value): Don't attempt to normalize if no value
given.
(_OptionalField): New helper class.
(optional_field): New.
(SymbolIndex.SCHEMA): Add `source' field.
(_WorkerContext): Add `use_compilation_database' attribute.
(_index_single_file): Use it.
(index_binary_directory): Set it.
* tests/fixture/Makefile (%.c.o):
(%.cpp.o): Compile with `-g' flag.
* tests/fixture/subdir/bar.cpp.o:
* tests/fixture/subdir/foo.c.o:
* tests/fixture/toplev.c.o: Recompile.
* tests/fixture/subdir/foo.c: Ensure one symbol is at expected
address.
* tests/test_cli.py (test_cli_indexing_with_compile_commands): Check
that `source' is set properly.
(test_cli_search_json_output): Update.
(test_cli_search_sexp_output): Update.
* tests/test_index.py (test_indexing): Update.
(test_indexing_adds_source_field_with_dwarfdump):
(test_indexing_adds_source_field_with_compilation_database): New tests.
(test_searching_by_address): Update.
* README.md: Add examples of searching by source path.

+++ tests/fixture/subdir/foo.c
@@ -1,6 +1,8 @@
-int foo;
-int foo_;
-int foo__;
+#include <stdint.h>
+
+uint64_t foo;
+uint64_t foo_;
+int foo__; /* Should be at address 0x10 */
 
 int uses_foo() { return foo; }
